+++ backend/internal/model/user.go
@@ -1,10 +1,10 @@
 package model
 
 type User struct {
-	ID       		int    `json:"id"`
-	TenantID   		int    `json:"tenant_id"`
-	Name       		string `json:"name"`
-	Email      		string `json:"email"`
-	PasswordHash 	string `json:"-"`
-	Role	   		string `json:"role"` // "agent" or "customer"
+	ID           int64  `json:"id"`
+	TenantID     int64  `json:"tenant_id"`
+	Name         string `json:"name"`
+	Email        string `json:"email"`
+	PasswordHash string `json:"-"`
+	Role         string `json:"role"` // "agent" or "customer"
 }
